Lots of changes
Lots of changes

Room size: increased 32 pixels in width and height
X and O: increased 32 pixels in width and height
Board: obj_corner and obj_section instead of just a few pieces of the
board
Checks: now if the place is empty, and doesn't touch the invisible
obj_position, it won't print anything, as intended
Draw: tiny adjustments to how the names are printed in the bottom
Board: clicking on the board now won't create the pieces on the board,
as stated in the previous bug fix, but now the spaces are more pixel
perfect

INTENDED "BUG": A pixel empty in between board and position possibility
is given, to avoid collision weirdness with mouse release over a frame
and then another frame being over the position given, making it look
like you clicked on the intersections and the pieces snapped to a
position. It's just how human eyes work.