Added scr_positionControl
This script is going to make the checks for pieces being used in a
certain position. In other words, this is how it checks for illegal
moves, and winning conditions.

For now it's just an initial check in one position, I'll research on a
better way of doing it instead of a myriad of ugly logical checks.