Bug Fix
Bug Fix:

The variable switching the position to taken was being activated with
mousing over the spot, instead of when button was released. Now fixed.

Change: array created to store the position values.

Possibly going to use integers instead of booleans in the variables so I
can use some calculation method to check for winning conditions.