Board Check
scr_positionControl controls the position of the pieces put on the
board. It checks the sets of pairs for the x and y axis, and iterates
count to 1 after every piece is added. It adds a count of one to the
variable count inside obj_manager, this is the variable that will check
for a victory status in the next script on the pipeline.