Victory Condition
This commit adds up Victory Condition.

If flag changes to true, using the mouse to try to place any other piece
won't work, as intended, after all, the match is over.

Right now, the winner is announced on the upper left below whose turn
time it is.

The game also announces if it's a tie. Such skill, much wow.