Bug Fix
I found a new bug on distanceControl. Pretty much the range didn't work
anymore as 32 pixels, since I increased the size of the board. I raised
the range to 48 and should be fine now.

I also made tiny changes to positionControl, the variables p0 to p8 were
being called under the logical statements twice, and I realized that it
was unnecessary. Now they are being switched before the logical
statement, which makes it a little faster.