Renaming functions, illegal move check, bug found
Illegal moves are now blocked from being done.

Rename the functions that had the prefix scr_, which is a GameMaker
convention. I always thought it was a silly one because if you have a
variable, it won't end with parenthesis in the end. So scr_function()
and function() can be clearly seen as functions even without the prefix.

Bug: the pixel I left empty in the corner or every spot unfortunately
isn't working as intended, and since I changed the range to check for
snapping positions, it might not be necessary anymore. Thus, I'm going
to remove it, and probably will solve the problem. That pixel on the
corner of every spot can be exploited to make illegal moves.