Bug Fix
Bug that allowed to create illegal moves should be solved. The collision
masks of the pieces were increased a bit, as well the corners and
sections of the board, in an attempt to overlap with the outer size
where moves aren't able to be placed, as well to cover the entirety of
the board, avoiding logical confusion.

Project is pretty much done, can use some polishing for Victory
statement, and offering to play it again, and make another little thing
or two, plus Git documentation.