Bug Fix Update
With the increase of the collision masks, it's needed to switch the
range again. It still won't fix the bug of a piece being put in one
position and snapping to the wrong one, but it will work on most times,
as long as it's not being tested at pixel perfect level on the corners.

To fix this, I'll need to write another function that checks mouse
position and obj_position to make sure that the piece won't snap to the
wrong one, and that could be some rework, so at first, this should work
as a project, and maybe maintaining the bug here is cool for the sake of
argument in the class.

It's not a hard fix, just a little work.